+++ packages/electron-esbuild/src/config/configurators/vite.ts
@@ -35,8 +35,8 @@ export class ViteConfigurator implements Configurator<TypeConfig.vite> {
         },
       },
       server: {
-        fsServe: {
-          root: '../..',
+        fs: {
+          allow: ['../..'],
         },
       },
     }