+++ packages/create-app/templates/react/esbuild.renderer.config.js
@@ -1,16 +1,11 @@
-const path = require('path')
+import path from 'path'
 
 /**
  * @var {Partial<import('esbuild').BuildOptions>}
  */
-module.exports = {
+export default {
   platform: 'browser',
   entryPoints: [path.resolve('src/renderer/index.jsx')],
   bundle: true,
   target: 'chrome89', // electron version target
-  loader: {
-    '.js': 'js',
-    '.jsx': 'jsx',
-    '.css': 'css',
-  },
 }