+++ packages/electron-esbuild/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## v3.0.0-rc.1
 
 ### Breaking Changes